refactor: move ecs related files to new `src/ecs` directory
