+++ src/ecs/command.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "common.h"
+#include "../common.h"
 #include "components.h"
 
 typedef struct
